+++ week4/assignment/src/types/routeType.ts
@@ -2,7 +2,7 @@ import { ReactNode } from 'react';
 
 export interface RouteType {
   path?: string;
-  index?: true;
+  index?: boolean;
   element: ReactNode;
-  children?: RouteType[]; // 중첩으로 라우팅
+  children?: RouteType[]; // 중첩으로 라우팅할 때 이렇게 사용
 }
